working to create a runner, reorged night writer class to be the shell and renamed the translator.
